Add Typing, Docstrings everywhere, organize the code a bit.
