+++ README.md
@@ -376,6 +376,10 @@ IMAGE_PROVIDER=sd
 HUGGINGFACE_API_TOKEN="YOUR_HUGGINGFACE_API_TOKEN"
 ```
 
+## Selenium
+
+sudo Xvfb :10 -ac -screen 0 1024x768x24 &
+DISPLAY=:10 your-client
 ## ⚠️ Limitations
 
 This experiment aims to showcase the potential of GPT-4 but comes with some limitations:
