+++ autogpt/agent/__init__.py
@@ -1,0 +1,4 @@
+from autogpt.agent.agent import Agent
+from autogpt.agent.agent_manager import AgentManager
+
+__all__ = ["Agent", "AgentManager"]
