+++ autogpt/chat.py
@@ -5,7 +5,7 @@ from openai.error import RateLimitError
 from autogpt import token_counter
 from autogpt.config import Config
 from autogpt.llm_utils import create_chat_completion
-from autogpt.logger import logger
+from autogpt.logs import logger
 
 cfg = Config()
