+++ autogpt/commands/file_operations.py
@@ -1,16 +1,29 @@
+"""File operations for AutoGPT"""
 import os
 import os.path
+from pathlib import Path
+from typing import Generator, List
 
 # Set a dedicated folder for file I/O
-working_directory = "auto_gpt_workspace"
+WORKING_DIRECTORY = Path(__file__).parent.parent / "auto_gpt_workspace"
 
 # Create the directory if it doesn't exist
-if not os.path.exists(working_directory):
-    os.makedirs(working_directory)
+if not os.path.exists(WORKING_DIRECTORY):
+    os.makedirs(WORKING_DIRECTORY)
+
+WORKING_DIRECTORY = str(WORKING_DIRECTORY)
 
 
-def safe_join(base, *paths):
-    """Join one or more path components intelligently."""
+def safe_join(base: str, *paths) -> str:
+    """Join one or more path components intelligently.
+
+    Args:
+        base (str): The base path
+        *paths (str): The paths to join to the base path
+
+    Returns:
+        str: The joined path
+    """
     new_path = os.path.join(base, *paths)
     norm_new_path = os.path.normpath(new_path)
 
@@ -20,7 +33,9 @@ def safe_join(base, *paths):
     return norm_new_path
 
 
-def split_file(content, max_length=4000, overlap=0):
+def split_file(
+    content: str, max_length: int = 4000, overlap: int = 0
+) -> Generator[str, None, None]:
     """
     Split text into chunks of a specified maximum length with a specified overlap
     between chunks.
@@ -45,10 +60,17 @@ def split_file(content, max_length=4000, overlap=0):
         start += max_length - overlap
 
 
-def read_file(filename) -> str:
-    """Read a file and return the contents"""
+def read_file(filename: str) -> str:
+    """Read a file and return the contents
+
+    Args:
+        filename (str): The name of the file to read
+
+    Returns:
+        str: The contents of the file
+    """
     try:
-        filepath = safe_join(working_directory, filename)
+        filepath = safe_join(WORKING_DIRECTORY, filename)
         with open(filepath, "r", encoding="utf-8") as f:
             content = f.read()
         return content
@@ -56,7 +78,9 @@ def read_file(filename) -> str:
         return f"Error: {str(e)}"
 
 
-def ingest_file(filename, memory, max_length=4000, overlap=200):
+def ingest_file(
+    filename: str, memory, max_length: int = 4000, overlap: int = 200
+) -> None:
     """
     Ingest a file by reading its content, splitting it into chunks with a specified
     maximum length and overlap, and adding the chunks to the memory storage.
@@ -88,10 +112,18 @@ def ingest_file(filename, memory, max_length=4000, overlap=200):
         print(f"Error while ingesting file '{filename}': {str(e)}")
 
 
-def write_to_file(filename, text):
-    """Write text to a file"""
+def write_to_file(filename: str, text: str) -> str:
+    """Write text to a file
+
+    Args:
+        filename (str): The name of the file to write to
+        text (str): The text to write to the file
+
+    Returns:
+        str: A message indicating success or failure
+    """
     try:
-        filepath = safe_join(working_directory, filename)
+        filepath = safe_join(WORKING_DIRECTORY, filename)
         directory = os.path.dirname(filepath)
         if not os.path.exists(directory):
             os.makedirs(directory)
@@ -99,43 +131,66 @@ def write_to_file(filename, text):
             f.write(text)
         return "File written to successfully."
     except Exception as e:
-        return "Error: " + str(e)
+        return f"Error: {str(e)}"
 
 
-def append_to_file(filename, text):
-    """Append text to a file"""
+def append_to_file(filename: str, text: str) -> str:
+    """Append text to a file
+
+    Args:
+        filename (str): The name of the file to append to
+        text (str): The text to append to the file
+
+    Returns:
+        str: A message indicating success or failure
+    """
     try:
-        filepath = safe_join(working_directory, filename)
+        filepath = safe_join(WORKING_DIRECTORY, filename)
         with open(filepath, "a") as f:
             f.write(text)
         return "Text appended successfully."
     except Exception as e:
-        return "Error: " + str(e)
+        return f"Error: {str(e)}"
 
 
-def delete_file(filename):
-    """Delete a file"""
+def delete_file(filename: str) -> str:
+    """Delete a file
+
+    Args:
+        filename (str): The name of the file to delete
+
+    Returns:
+        str: A message indicating success or failure
+    """
     try:
-        filepath = safe_join(working_directory, filename)
+        filepath = safe_join(WORKING_DIRECTORY, filename)
         os.remove(filepath)
         return "File deleted successfully."
     except Exception as e:
-        return "Error: " + str(e)
+        return f"Error: {str(e)}"
 
 
-def search_files(directory):
+def search_files(directory: str) -> List[str]:
+    """Search for files in a directory
+
+    Args:
+        directory (str): The directory to search in
+
+    Returns:
+        List[str]: A list of files found in the directory
+    """
     found_files = []
 
-    if directory == "" or directory == "/":
-        search_directory = working_directory
+    if directory in {"", "/"}:
+        search_directory = WORKING_DIRECTORY
     else:
-        search_directory = safe_join(working_directory, directory)
+        search_directory = safe_join(WORKING_DIRECTORY, directory)
 
     for root, _, files in os.walk(search_directory):
         for file in files:
             if file.startswith("."):
                 continue
-            relative_path = os.path.relpath(os.path.join(root, file), working_directory)
+            relative_path = os.path.relpath(os.path.join(root, file), WORKING_DIRECTORY)
             found_files.append(relative_path)
 
     return found_files
