+++ autogpt/config/__init__.py
@@ -1,0 +1,14 @@
+"""
+This module contains the configuration classes for AutoGPT.
+"""
+from autogpt.config.ai_config import AIConfig
+from autogpt.config.config import check_openai_api_key, Config
+from autogpt.config.singleton import AbstractSingleton, Singleton
+
+__all__ = [
+    "check_openai_api_key",
+    "AbstractSingleton",
+    "AIConfig",
+    "Config",
+    "Singleton",
+]
