+++ autogpt/config/config.py
@@ -1,30 +1,15 @@
-import abc
+"""Configuration class to store the state of bools for different scripts access."""
 import os
+from colorama import Fore
+
+from autogpt.config.singleton import Singleton
 
 import openai
 import yaml
+
 from dotenv import load_dotenv
 
-# Load environment variables from .env file
-load_dotenv()
-
-
-class Singleton(abc.ABCMeta, type):
-    """
-    Singleton metaclass for ensuring only one instance of a class.
-    """
-
-    _instances = {}
-
-    def __call__(cls, *args, **kwargs):
-        """Call method for the singleton metaclass."""
-        if cls not in cls._instances:
-            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
-        return cls._instances[cls]
-
-
-class AbstractSingleton(abc.ABC, metaclass=Singleton):
-    pass
+load_dotenv(verbose=True)
 
 
 class Config(metaclass=Singleton):
@@ -32,7 +17,7 @@ class Config(metaclass=Singleton):
     Configuration class to store the state of bools for different scripts access.
     """
 
-    def __init__(self):
+    def __init__(self) -> None:
         """Initialize the Config class"""
         self.debug_mode = False
         self.continuous_mode = False
@@ -81,10 +66,12 @@ class Config(metaclass=Singleton):
         self.huggingface_api_token = os.getenv("HUGGINGFACE_API_TOKEN")
 
         # User agent headers to use when browsing web
-        # Some websites might just completely deny request with an error code if no user agent was found.
+        # Some websites might just completely deny request with an error code if
+        # no user agent was found.
         self.user_agent = os.getenv(
             "USER_AGENT",
-            "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_4) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/83.0.4103.97 Safari/537.36",
+            "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_4) AppleWebKit/537.36"
+            " (KHTML, like Gecko) Chrome/83.0.4103.97 Safari/537.36",
         )
         self.redis_host = os.getenv("REDIS_HOST", "localhost")
         self.redis_port = os.getenv("REDIS_PORT", "6379")
@@ -108,15 +95,17 @@ class Config(metaclass=Singleton):
             The matching deployment id if found, otherwise an empty string.
         """
         if model == self.fast_llm_model:
-            return self.azure_model_to_deployment_id_map["fast_llm_model_deployment_id"]
+            return self.azure_model_to_deployment_id_map[
+                "fast_llm_model_deployment_id"
+            ]  # type: ignore
         elif model == self.smart_llm_model:
             return self.azure_model_to_deployment_id_map[
                 "smart_llm_model_deployment_id"
-            ]
+            ]  # type: ignore
         elif model == "text-embedding-ada-002":
             return self.azure_model_to_deployment_id_map[
                 "embedding_model_deployment_id"
-            ]
+            ]  # type: ignore
         else:
             return ""
 
@@ -124,7 +113,8 @@ class Config(metaclass=Singleton):
 
     def load_azure_config(self, config_file: str = AZURE_CONFIG_FILE) -> None:
         """
-        Loads the configuration parameters for Azure hosting from the specified file path as a yaml file.
+        Loads the configuration parameters for Azure hosting from the specified file
+          path as a yaml file.
 
         Parameters:
             config_file(str): The path to the config yaml file. DEFAULT: "../azure.yaml"
@@ -148,74 +138,86 @@ class Config(metaclass=Singleton):
         )
         self.azure_model_to_deployment_id_map = config_params.get("azure_model_map", [])
 
-    def set_continuous_mode(self, value: bool):
+    def set_continuous_mode(self, value: bool) -> None:
         """Set the continuous mode value."""
         self.continuous_mode = value
 
-    def set_continuous_limit(self, value: int):
+    def set_continuous_limit(self, value: int) -> None:
         """Set the continuous limit value."""
         self.continuous_limit = value
 
-    def set_speak_mode(self, value: bool):
+    def set_speak_mode(self, value: bool) -> None:
         """Set the speak mode value."""
         self.speak_mode = value
 
-    def set_fast_llm_model(self, value: str):
+    def set_fast_llm_model(self, value: str) -> None:
         """Set the fast LLM model value."""
         self.fast_llm_model = value
 
-    def set_smart_llm_model(self, value: str):
+    def set_smart_llm_model(self, value: str) -> None:
         """Set the smart LLM model value."""
         self.smart_llm_model = value
 
-    def set_fast_token_limit(self, value: int):
+    def set_fast_token_limit(self, value: int) -> None:
         """Set the fast token limit value."""
         self.fast_token_limit = value
 
-    def set_smart_token_limit(self, value: int):
+    def set_smart_token_limit(self, value: int) -> None:
         """Set the smart token limit value."""
         self.smart_token_limit = value
 
-    def set_browse_chunk_max_length(self, value: int):
+    def set_browse_chunk_max_length(self, value: int) -> None:
         """Set the browse_website command chunk max length value."""
         self.browse_chunk_max_length = value
 
-    def set_browse_summary_max_token(self, value: int):
+    def set_browse_summary_max_token(self, value: int) -> None:
         """Set the browse_website command summary max token value."""
         self.browse_summary_max_token = value
 
-    def set_openai_api_key(self, value: str):
+    def set_openai_api_key(self, value: str) -> None:
         """Set the OpenAI API key value."""
         self.openai_api_key = value
 
-    def set_elevenlabs_api_key(self, value: str):
+    def set_elevenlabs_api_key(self, value: str) -> None:
         """Set the ElevenLabs API key value."""
         self.elevenlabs_api_key = value
 
-    def set_elevenlabs_voice_1_id(self, value: str):
+    def set_elevenlabs_voice_1_id(self, value: str) -> None:
         """Set the ElevenLabs Voice 1 ID value."""
         self.elevenlabs_voice_1_id = value
 
-    def set_elevenlabs_voice_2_id(self, value: str):
+    def set_elevenlabs_voice_2_id(self, value: str) -> None:
         """Set the ElevenLabs Voice 2 ID value."""
         self.elevenlabs_voice_2_id = value
 
-    def set_google_api_key(self, value: str):
+    def set_google_api_key(self, value: str) -> None:
         """Set the Google API key value."""
         self.google_api_key = value
 
-    def set_custom_search_engine_id(self, value: str):
+    def set_custom_search_engine_id(self, value: str) -> None:
         """Set the custom search engine id value."""
         self.custom_search_engine_id = value
 
-    def set_pinecone_api_key(self, value: str):
+    def set_pinecone_api_key(self, value: str) -> None:
         """Set the Pinecone API key value."""
         self.pinecone_api_key = value
 
-    def set_pinecone_region(self, value: str):
+    def set_pinecone_region(self, value: str) -> None:
         """Set the Pinecone region value."""
         self.pinecone_region = value
 
-    def set_debug_mode(self, value: bool):
+    def set_debug_mode(self, value: bool) -> None:
         """Set the debug mode value."""
         self.debug_mode = value
+
+
+def check_openai_api_key() -> None:
+    """Check if the OpenAI API key is set in config.py or as an environment variable."""
+    cfg = Config()
+    if not cfg.openai_api_key:
+        print(
+            Fore.RED
+            + "Please set your OpenAI API key in .env or as an environment variable."
+        )
+        print("You can get your key from https://beta.openai.com/account/api-keys")
+        exit(1)
