+++ autogpt/json_fixes/auto_fix.py
@@ -1,0 +1,40 @@
+"""This module contains the function to fix JSON strings using GPT-3."""
+import json
+from autogpt.llm_utils import call_ai_function
+from autogpt.logs import logger
+
+
+def fix_json(json_str: str, schema: str) -> str:
+    """Fix the given JSON string to make it parseable and fully compliant with the provided schema."""
+    # Try to fix the JSON using GPT:
+    function_string = "def fix_json(json_str: str, schema:str=None) -> str:"
+    args = [f"'''{json_str}'''", f"'''{schema}'''"]
+    description_string = (
+        "Fixes the provided JSON string to make it parseable"
+        " and fully compliant with the provided schema.\n If an object or"
+        " field specified in the schema isn't contained within the correct"
+        " JSON, it is omitted.\n This function is brilliant at guessing"
+        " when the format is incorrect."
+    )
+
+    # If it doesn't already start with a "`", add one:
+    if not json_str.startswith("`"):
+        json_str = "```json\n" + json_str + "\n```"
+    result_string = call_ai_function(
+        function_string, args, description_string, model=cfg.fast_llm_model
+    )
+    logger.debug("------------ JSON FIX ATTEMPT ---------------")
+    logger.debug(f"Original JSON: {json_str}")
+    logger.debug("-----------")
+    logger.debug(f"Fixed JSON: {result_string}")
+    logger.debug("----------- END OF FIX ATTEMPT ----------------")
+
+    try:
+        json.loads(result_string)  # just check the validity
+        return result_string
+    except:  # noqa: E722
+        # Get the call stack:
+        # import traceback
+        # call_stack = traceback.format_exc()
+        # print(f"Failed to fix JSON: '{json_str}' "+call_stack)
+        return "failed"
