+++ autogpt/memory/pinecone.py
@@ -1,7 +1,7 @@
 import pinecone
 from colorama import Fore, Style
 
-from autogpt.logger import logger
+from autogpt.logs import logger
 from autogpt.memory.base import MemoryProviderSingleton, get_ada_embedding
 
 
@@ -16,7 +16,8 @@ class PineconeMemory(MemoryProviderSingleton):
         table_name = "auto-gpt"
         # this assumes we don't start with memory.
         # for now this works.
-        # we'll need a more complicated and robust system if we want to start with memory.
+        # we'll need a more complicated and robust system if we want to start with
+        #  memory.
         self.vec_num = 0
 
         try:
@@ -28,8 +29,10 @@ class PineconeMemory(MemoryProviderSingleton):
                 Style.BRIGHT + str(e) + Style.RESET_ALL,
             )
             logger.double_check(
-                "Please ensure you have setup and configured Pinecone properly for use. "
-                + f"You can check out {Fore.CYAN + Style.BRIGHT}https://github.com/Torantulino/Auto-GPT#-pinecone-api-key-setup{Style.RESET_ALL} to ensure you've set up everything correctly."
+                "Please ensure you have setup and configured Pinecone properly for use."
+                + f"You can check out {Fore.CYAN + Style.BRIGHT}"
+                "https://github.com/Torantulino/Auto-GPT#-pinecone-api-key-setup"
+                f"{Style.RESET_ALL} to ensure you've set up everything correctly."
             )
             exit(1)
 
@@ -42,7 +45,7 @@ class PineconeMemory(MemoryProviderSingleton):
     def add(self, data):
         vector = get_ada_embedding(data)
         # no metadata here. We may wish to change that long term.
-        resp = self.index.upsert([(str(self.vec_num), vector, {"raw_text": data})])
+        self.index.upsert([(str(self.vec_num), vector, {"raw_text": data})])
         _text = f"Inserting data into memory at index: {self.vec_num}:\n data: {data}"
         self.vec_num += 1
         return _text
