+++ autogpt/speech/gtts.py
@@ -1,0 +1,21 @@
+""" GTTS Voice. """
+import os
+from playsound import playsound
+import gtts
+
+from autogpt.speech.base import VoiceBase
+
+
+class GTTSVoice(VoiceBase):
+    """GTTS Voice."""
+
+    def _setup(self) -> None:
+        pass
+
+    def _speech(self, text: str, _: int = 0) -> bool:
+        """Play the given text."""
+        tts = gtts.gTTS(text)
+        tts.save("speech.mp3")
+        playsound("speech.mp3", True)
+        os.remove("speech.mp3")
+        return True
