+++ autogpt/spinner.py
@@ -1,3 +1,4 @@
+"""A simple spinner module"""
 import itertools
 import sys
 import threading
@@ -7,15 +8,20 @@ import time
 class Spinner:
     """A simple spinner class"""
 
-    def __init__(self, message="Loading...", delay=0.1):
-        """Initialize the spinner class"""
+    def __init__(self, message: str = "Loading...", delay: float = 0.1) -> None:
+        """Initialize the spinner class
+
+        Args:
+            message (str): The message to display.
+            delay (float): The delay between each spinner update.
+        """
         self.spinner = itertools.cycle(["-", "/", "|", "\\"])
         self.delay = delay
         self.message = message
         self.running = False
         self.spinner_thread = None
 
-    def spin(self):
+    def spin(self) -> None:
         """Spin the spinner"""
         while self.running:
             sys.stdout.write(f"{next(self.spinner)} {self.message}\r")
@@ -23,14 +29,20 @@ class Spinner:
             time.sleep(self.delay)
             sys.stdout.write(f"\r{' ' * (len(self.message) + 2)}\r")
 
-    def __enter__(self):
+    def __enter__(self) -> None:
         """Start the spinner"""
         self.running = True
         self.spinner_thread = threading.Thread(target=self.spin)
         self.spinner_thread.start()
 
-    def __exit__(self, exc_type, exc_value, exc_traceback):
-        """Stop the spinner"""
+    def __exit__(self, exc_type, exc_value, exc_traceback) -> None:
+        """Stop the spinner
+
+        Args:
+            exc_type (Exception): The exception type.
+            exc_value (Exception): The exception value.
+            exc_traceback (Exception): The exception traceback.
+        """
         self.running = False
         if self.spinner_thread is not None:
             self.spinner_thread.join()
