+++ tests/unit/test_browse_scrape_text.py
@@ -2,7 +2,7 @@
 
 import requests
 
-from autogpt.browse import scrape_text
+from autogpt.commands.web_requests import scrape_text
 
 """
 Code Analysis
